Part 3_1 finished
The 10logcoding is better. I have add one variable to the function encoder to identify if it is length.

I am doing the part 3_2. I am just going to tell that we can do compile the length of each sequence of 0 and 1 and then do a shannon encoding or a huffman encoding.